Moved the Covid Tooltip to its own file
