+++ src/_scss/pages/agencyV2/overview/_overview.scss
@@ -29,7 +29,7 @@
                 text-align: center;
                 height: rem(25);
                 border-radius: rem(4);
-                width: rem(175);
+                padding: 0 .8rem;
                 margin-top: 1.1rem;
                 .tooltip-spacer {
                     display: none;
